LayoutTests/imported/w3c/web-platform-tests/cookie-store/cookieStore_get_set_across_origins.sub.https-expected.txt


FAIL cookieStore.get() sees cookieStore.set() in cross-origin frame assert_equals: expected 1 but got 0
FAIL cookieStore.get() in cross-origin frame sees cookieStore.set() assert_not_equals: got disallowed value null
